+++ sdk/eventhub/arm-eventhub/src/models/parameters.ts
@@ -22,8 +22,7 @@ import {
   PrivateEndpointConnection as PrivateEndpointConnectionMapper,
   Eventhub as EventhubMapper,
   ArmDisasterRecovery as ArmDisasterRecoveryMapper,
-  ConsumerGroup as ConsumerGroupMapper,
-  SchemaGroup as SchemaGroupMapper
+  ConsumerGroup as ConsumerGroupMapper
 } from "../models/mappers";
 
 export const accept: OperationParameter = {
@@ -64,7 +63,7 @@ export const subscriptionId: OperationURLParameter = {
 export const apiVersion: OperationQueryParameter = {
   parameterPath: "apiVersion",
   mapper: {
-    defaultValue: "2021-11-01",
+    defaultValue: "2021-06-01-preview",
     isConstant: true,
     serializedName: "api-version",
     type: {
@@ -288,26 +287,6 @@ export const consumerGroupName: OperationURLParameter = {
       MinLength: 1
     },
     serializedName: "consumerGroupName",
-    required: true,
-    type: {
-      name: "String"
-    }
-  }
-};
-
-export const parameters11: OperationParameter = {
-  parameterPath: "parameters",
-  mapper: SchemaGroupMapper
-};
-
-export const schemaGroupName: OperationURLParameter = {
-  parameterPath: "schemaGroupName",
-  mapper: {
-    constraints: {
-      MaxLength: 256,
-      MinLength: 1
-    },
-    serializedName: "schemaGroupName",
     required: true,
     type: {
       name: "String"