+++ sdk/eventhub/arm-eventhub/src/operations/privateEndpointConnections.ts
@@ -208,12 +208,10 @@ export class PrivateEndpointConnectionsImpl
       },
       deleteOperationSpec
     );
-    const poller = new LroEngine(lro, {
+    return new LroEngine(lro, {
       resumeFrom: options?.resumeFrom,
       intervalInMs: options?.updateIntervalInMs
     });
-    await poller.poll();
-    return poller;
   }
 
   /**